requirements.txt
Django==2.2.14
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5
